fix(SetEnv): escape backslashes in PATH
The SetEnv() utility function uses Vim's eval() to set the value of an environment variable. This fails with variable values that contain backslashes (such as Windows directories in PATH) because the backslashes are treated as character escapes.

This (partial) fix replaces the double quotes with single quotes to prevent the backslashes from being interpreted as escapes. This is enough to fix my installation problems with `:GoInstallBinaries` but is not a complete solution because it does not escape single quotes (`'` should be `''`).